# fix: improve performance by parallelizing build, browser, and preview