These changes aim to improve the plugin's performance. 

## Changes

- Spawns Puppeteer as soon as the first OG image route is detected during the route modules transformation (vs waiting for the entire build to finish, _then_ spawning PT).
- Spawns Vite preview server as soon as the first OG image route is detected. 
- Drops `.toBuffer()` and preserves the `sharp.Sharp` stream, which is `Duplex`. Passes it directly to `fs.promises.writeFile()` so we don't read it twice (we don't really need to read it at all).
- Merges DOM element bounding box calculation and scrolling it to viewport into a single `page.evaluage()`.

## Profiling

```
loader: ~4s
pageload: ~1s
screenshot: ~4s
```

Maybe I'm measuring the screenshot time wrong but it is exponentially growing with each screenshot. This may indicate a memory leak. 

## Saved time

**~3s**.

## Possible time-saves

- [x] ~~If the app is using Single fetch in Remix, fetch all the OG image routes resources as a single `fetch` request. This should be a significant time-save.~~
  - My understanding on how `_routes` works in Remix was wrong. It's meant to fetch loader data for _matching_ loaders, not a list of arbitrary loaders. If you provide it a list of random loaders, Remix will ignore those that have nothing to do with the current route (the fetching is also route specific so you cannot fetch all on root).
- [x] ~~Skip closing pages?~~ The plugin will close the entire browser, which will close all the pages. This may negatively impact the RAM, slowing down the entire automation. 
- [x] ~~Consider [ignoring certain network requests](https://www.roborabbit.com/blog/mastering-puppeteer-tips-and-tricks-for-effortless-web-automation/#tip-4-disable-css-images-and-other-unneeded-resources-to-speed-up-puppeteer) to make the page load faster. This will only have effect if your nested layout is loading general data, like ads, that has no effect on the image.~~
  - This can be a nice customization option, but for now this results in no performance gain as I'm setting a simple image generation as a benchmark. 
- [x] ~~Consider chaining image writes to disk right in `generateRouteImages` vs adding it in `.then()`.~~
  - This yields no performance difference.