@kettanaito 2024-09-02
## Released: v0.3.4 🎉

This has been released in v0.3.4!

- 📄 [**Release notes**](https://github.com/kettanaito/remix-og-image/releases/tag/v0.3.4)
- 📦 [npm package](https://www.npmjs.com/package/remix-og-image/v/0.3.4)

Make sure to always update to the latest version (`npm i remix-og-image@latest`) to get the newest features and bug fixes.

---

_Predictable release automation by [@ossjs/release](https://github.com/ossjs/release)_.